version_3

+++ src/main/java/com/example/calculatorTest/calculatorTest/service/CalculatorService.java
@@ -4,11 +4,11 @@ public interface CalculatorService {
 
     String helloCalculator();
 
-    int addTwoNumbers(int num1, int num2);
+    long addTwoNumbers(int num1, int num2);
 
-    int subtractTwoNumbers(int num1, int num2);
+    long subtractTwoNumbers(int num1, int num2);
 
-    int multiplyTwoNumbers(int num1, int num2);
+    long multiplyTwoNumbers(int num1, int num2);
 
     double divideTwoNumbers(int num1, int num2);
